Compilation fixes and move to unsigned long for offsets

+++ DumpSampleFiles.cpp
@@ -63,7 +63,7 @@ void dumpGraph(
 
     if(weights) {
       f << " weight='";
-      if(!isnan((*weights)[i]))
+      if(!isnan(static_cast<long double>((*weights)[i])))
         f << (*weights)[i];
       f << "'";
     }
